created customviews. moved implementation to statefragment
