+++ app/src/main/java/org/oppia/app/fragment/FragmentComponent.kt
@@ -7,7 +7,6 @@ import org.oppia.app.home.HomeFragment
 import org.oppia.app.player.exploration.ExplorationFragment
 import org.oppia.app.player.state.StateFragment
 import org.oppia.app.player.audio.AudioFragment
-import org.oppia.app.player.content_interaction.ContentInteractionFragment
 import org.oppia.app.topic.conceptcard.ConceptCardFragment
 
 /** Root subcomponent for all fragments. */
@@ -25,5 +24,4 @@ interface FragmentComponent {
   fun inject(homeFragment: HomeFragment)
   fun inject(stateFragment: StateFragment)
   fun inject(conceptCardFragment: ConceptCardFragment)
-  fun inject(contentInteractionFragment: ContentInteractionFragment)
 }
